fix(decompress): ReadLoads returns ssize_t, no false error on >2GB cores
B60 (regression from 608fd44): ReadLoads returned int, but the value is the
total uncompressed LOADS bytes written. For a process with more than 2GiB of
readable mappings, loads_size truncated to a negative int, so the B54
`rc < 0` check rejected a perfectly valid dump. Reproduced with a 3.2GiB
anonymous-mapping process: dump succeeded (13.6MB acore), decompress failed
"read loads failed, core incomplete" while having already written the full
3.2GB core.

Change ReadLoads to return ssize_t (0..2^63-1 for valid loads, -1 on
error) and check that in decompress.

Verified: the 3.2GiB acore now decompresses to a gdb-readable 3.2GB core
with correct memory; the ASAN build had to be rebuilt explicitly because
`make clean` drops it (documented in the round notes).

+++ src/core.h
@@ -292,7 +292,9 @@ private:
     // read acore 
     int VerifyFileHeader(Lz4Stream& in);
     int ReadMeta(Lz4Stream& in);
-    int ReadLoads(Lz4Stream& in, FILE* fout);
+    // 返回实际写出的未压缩字节数（可为数 GB，须用 ssize_t，不能用 int——
+    // >2GB 会截断成负数被调用方误判为错误）；-1 表示读失败。
+    ssize_t ReadLoads(Lz4Stream& in, FILE* fout);
     int ReadElfHeader(Lz4Stream& in);
  
     // write core
